[CCUBE-2115][GZ] move back tabs to storybook commmon top level
