+++ stories/storybook-common/index.ts
@@ -2,5 +2,5 @@ export * from "./api-table";
 export * from "./decorators";
 export * from "./docs";
 export * from "./inspectors";
-export * from "./navigation";
+export * from "./tabs";
 export * from "./token-table";
